% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiverse.workshop-package.R
\docType{package}
\name{epiverse.workshop-package}
\alias{epiverse.workshop}
\alias{epiverse.workshop-package}
\title{epiverse.workshop: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BenSinger01/epiverse.workshop}
  \item Report bugs at \url{https://github.com/BenSinger01/epiverse.workshop/issues}
}

}
\author{
\strong{Maintainer}: Ben Singer \email{benjaminjohnsinger@gmail.com} (\href{https://orcid.org/0000-0003-3062-2620}{ORCID}) [copyright holder]

}
\keyword{internal}
